add reservation to tnodb mock
+ update http provision source to match tnodb change
