+++ tools/tnodb_mock/tnodb_test.go
@@ -5,7 +5,6 @@ import (
 	"net"
 	"testing"
 
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
@@ -23,42 +22,5 @@ func TestPrefixAllocation(t *testing.T) {
 		subnet, err := allocate(alloc)
 		require.NoError(t, err)
 		fmt.Println(subnet.String())
-	}
-}
-
-func TestNetIP(t *testing.T) {
-	for _, tc := range []struct {
-		netZero   string
-		alloc     string
-		allocSize int
-		result    string
-	}{
-		{
-			netZero:   "2a02:2788:cc02::",
-			alloc:     "2a02:2788:cc02:abcd::",
-			allocSize: 48,
-			result:    "2a02:2788:cc02::abcd",
-		},
-		{
-			allocSize: 32,
-			netZero:   "2a02:2788::",
-			alloc:     "2a02:2788:cc02:abcd::",
-			result:    "2a02:2788::cc02:abcd",
-		},
-	} {
-
-		t.Run(fmt.Sprintf("%d", tc.allocSize), func(t *testing.T) {
-			netZero := &net.IPNet{
-				IP:   net.ParseIP(tc.netZero),
-				Mask: net.CIDRMask(64, 128),
-			}
-			alloc := &net.IPNet{
-				IP:   net.ParseIP(tc.alloc),
-				Mask: net.CIDRMask(64, 128),
-			}
-
-			ip := netZeroIP(netZero, alloc, tc.allocSize)
-			assert.Equal(t, net.ParseIP(tc.result), ip)
-		})
 	}
 }
